// client-side/src/api/types/paymentHistoryResponseStatus.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * payment-systems API
 * OpenAPI spec version: 0.0.1
 */

/**
 * Transaction status
 */
export type PaymentHistoryResponseStatus = typeof PaymentHistoryResponseStatus[keyof typeof PaymentHistoryResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentHistoryResponseStatus = {
  PENDING: 'PENDING',
  SUCCEEDED: 'SUCCEEDED',
  CANCELLED: 'CANCELLED',
  FAILED: 'FAILED',
} as const;