// client-side/src/api/types/subscriptionResponseStatus.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * payment-systems API
 * OpenAPI spec version: 0.0.1
 */

/**
 * Current subscription status
 */
export type SubscriptionResponseStatus = typeof SubscriptionResponseStatus[keyof typeof SubscriptionResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SubscriptionResponseStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
} as const;